refactor(anomaly detection): extract shared logic to utils
